refactor: integrate stdout collector into uplink
