+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
 exclude = [
     "tools/deserialize-backup",
     "tools/simulator",
-    "tools/stdout-collector",
     "tools/tunshell",
     "tools/utils",
 ]
