--- tools/stdout-collector/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "stdout-collector"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-futures-util = { version = "0.3", features = ["sink"] }
-log = "0.4"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1.0"
-structopt = "0.3"
-thiserror = "1"
-tokio = { version = "1", features = ["full"] }
-tokio-util = { version = "0.7", features = ["codec", "time"] }
