+++ uplink/src/base/mod.rs
@@ -191,6 +191,11 @@ impl Default for DeviceShadowConfig {
     }
 }
 
+#[derive(Clone, Debug, Deserialize)]
+pub struct StdoutConfig {
+    pub stream_name: String,
+}
+
 #[derive(Debug, Clone, Deserialize, Default)]
 pub struct Config {
     pub project_id: String,
@@ -218,6 +223,7 @@ pub struct Config {
     pub system_stats: Stats,
     pub simulator: Option<SimulatorConfig>,
     pub ota_installer: Option<InstallerConfig>,
+    pub stdout: Option<StdoutConfig>,
     #[serde(default)]
     pub device_shadow: DeviceShadowConfig,
     #[serde(default)]
